backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.6.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.10
langchain-anthropic>=0.0.1
langchain-chroma>=0.0.3
langchain-community>=0.0.10
chromadb>=0.4.22
openai>=1.12.0
anthropic>=0.18.1
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.1
markdown>=3.5.2 